+++ fabric/src/main/java/io/github/flemmli97/improvedmobs/fabric/config/ConfigSpecs.java
@@ -6,42 +6,50 @@ import io.github.flemmli97.improvedmobs.config.Config;
 import io.github.flemmli97.improvedmobs.config.EntityItemConfig;
 import io.github.flemmli97.improvedmobs.config.EntityModifyFlagConfig;
 import io.github.flemmli97.improvedmobs.config.MobClassMapConfig;
+import io.github.flemmli97.tenshilib.common.config.CommentedJsonConfig;
 import io.github.flemmli97.tenshilib.common.config.JsonConfig;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.ChatFormatting;
+import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public final class ConfigSpecs {
 
-    public static JsonConfig<ClientConfigVals> clientConfig;
-    public static JsonConfig<CommonConfigVals> commonConfig;
+    public static JsonConfig<CommentedJsonConfig> clientConfig;
+    public static ClientConfigVals clientConfigSpecs;
+
+    public static JsonConfig<CommentedJsonConfig> commonConfig;
+    public static CommonConfigVals commonConfigSpecs;
 
     public static void initClientConfig() {
-        clientConfig = new JsonConfig<>(FabricLoader.getInstance().getConfigDir().resolve("improvedmobs").resolve("client.json").toFile(),
-                ConfigSpecs.ClientConfigVals.class, new ConfigSpecs.ClientConfigVals());
+        Pair<JsonConfig<CommentedJsonConfig>, ClientConfigVals> pair = CommentedJsonConfig.Builder
+                .create(FabricLoader.getInstance().getConfigDir().resolve("improvedmobs").resolve("client.json"), 1, ClientConfigVals::new);
+        clientConfig = pair.getKey();
+        clientConfigSpecs = pair.getValue();
     }
 
     public static void initCommonConfig() {
-        commonConfig = new JsonConfig<>(FabricLoader.getInstance().getConfigDir().resolve("improvedmobs").resolve("common.json").toFile(),
-                ConfigSpecs.CommonConfigVals.class, new ConfigSpecs.CommonConfigVals());
+        Pair<JsonConfig<CommentedJsonConfig>, CommonConfigVals> pair = CommentedJsonConfig.Builder
+                .create(FabricLoader.getInstance().getConfigDir().resolve("improvedmobs").resolve("client.json"), 1, CommonConfigVals::new);
+        commonConfig = pair.getKey();
+        commonConfigSpecs = pair.getValue();
     }
 
     public final static class ClientConfigVals {
 
-        public final CommentedVal<Integer> guiX;
-        public final CommentedVal<Integer> guiY;
-        public final CommentedVal<ChatFormatting> color;
-        public final CommentedVal<Double> scale;
-        public final CommentedVal<Boolean> showDifficulty;
+        public final CommentedJsonConfig.IntVal guiX;
+        public final CommentedJsonConfig.IntVal guiY;
+        public final CommentedJsonConfig.CommentedVal<ChatFormatting> color;
+        public final CommentedJsonConfig.DoubleVal scale;
+        public final CommentedJsonConfig.CommentedVal<Boolean> showDifficulty;
 
-        public ClientConfigVals() {
-            CommentedVal.Builder builder = new CommentedVal.Builder();
-            this.guiX = builder.define("Gui X", 5);
-            this.guiY = builder.define("Gui Y", 5);
+        public ClientConfigVals(CommentedJsonConfig.Builder builder) {
+            this.guiX = builder.defineInRange("Gui X", 5, 0, Integer.MAX_VALUE);
+            this.guiY = builder.defineInRange("Gui Y", 5, 0, Integer.MAX_VALUE);
             this.color = builder.comment("Textformatting codes for the display of the difficulty").define("Difficulty color", ChatFormatting.DARK_PURPLE);
-            this.scale = builder.comment("Scaling of the difficulty text").define("Text Scale", 1D);
+            this.scale = builder.comment("Scaling of the difficulty text").defineInRange("Text Scale", 1D, 0, Double.MAX_VALUE);
             this.showDifficulty = builder.comment("Show the the difficulty text").define("Show Difficulty", true);
         }
     }
@@ -49,93 +57,92 @@ public final class ConfigSpecs {
     public final static class CommonConfigVals {
 
         //General
-        public final CommentedVal<Boolean> enableDifficultyScaling;
-        public final CommentedVal<Integer> difficultyDelay;
-        public final CommentedVal<List<String>> increaseHandler;
-        public final CommentedVal<Boolean> ignorePlayers;
-        public final CommentedVal<Boolean> shouldPunishTimeSkip;
-        public final CommentedVal<Boolean> friendlyFire;
-        public final CommentedVal<List<String>> petArmorBlackList;
-        public final CommentedVal<Boolean> petWhiteList;
-        public final CommentedVal<Boolean> doIMDifficulty;
-        public final CommentedVal<Config.DifficultyType> difficultyType;
+        public final CommentedJsonConfig.CommentedVal<Boolean> enableDifficultyScaling;
+        public final CommentedJsonConfig.IntVal difficultyDelay;
+        public final CommentedJsonConfig.CommentedVal<List<String>> increaseHandler;
+        public final CommentedJsonConfig.CommentedVal<Boolean> ignorePlayers;
+        public final CommentedJsonConfig.CommentedVal<Boolean> shouldPunishTimeSkip;
+        public final CommentedJsonConfig.CommentedVal<Boolean> friendlyFire;
+        public final CommentedJsonConfig.CommentedVal<List<String>> petArmorBlackList;
+        public final CommentedJsonConfig.CommentedVal<Boolean> petWhiteList;
+        public final CommentedJsonConfig.CommentedVal<Boolean> doIMDifficulty;
+        public final CommentedJsonConfig.CommentedVal<Config.DifficultyType> difficultyType;
 
         //Black-WhiteList
-        public final CommentedVal<List<String>> flagBlacklist;
-        public final CommentedVal<List<String>> entityBlacklist;
-        public final CommentedVal<Boolean> mobAttributeWhitelist;
-        public final CommentedVal<Boolean> armorMobWhitelist;
-        public final CommentedVal<Boolean> heldMobWhitelist;
-        public final CommentedVal<Boolean> mobListBreakWhitelist;
-        public final CommentedVal<Boolean> mobListUseWhitelist;
-        public final CommentedVal<Boolean> mobListLadderWhitelist;
-        public final CommentedVal<Boolean> mobListStealWhitelist;
-        public final CommentedVal<Boolean> mobListBoatWhitelist;
-        public final CommentedVal<Boolean> mobListFlyWhitelist;
-        public final CommentedVal<Boolean> targetVillagerWhitelist;
-        public final CommentedVal<Boolean> neutralAggroWhitelist;
+        public final CommentedJsonConfig.CommentedVal<List<String>> flagBlacklist;
+        public final CommentedJsonConfig.CommentedVal<List<String>> entityBlacklist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobAttributeWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> armorMobWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> heldMobWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListBreakWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListUseWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListLadderWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListStealWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListBoatWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> mobListFlyWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> targetVillagerWhitelist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> neutralAggroWhitelist;
 
         //Debug
-        public final CommentedVal<Boolean> debugPath;
+        public final CommentedJsonConfig.CommentedVal<Boolean> debugPath;
 
         //Integration
-        public final CommentedVal<Boolean> useScalingHealthMod;
+        public final CommentedJsonConfig.CommentedVal<Boolean> useScalingHealthMod;
 
         //AI
-        public final CommentedVal<List<String>> breakableBlocks;
-        public final CommentedVal<Boolean> breakingAsBlacklist;
-        public final CommentedVal<Boolean> useBlockBreakSound;
-        public final CommentedVal<Double> breakerChance;
-        public final CommentedVal<Integer> breakerInitCooldown;
-        public final CommentedVal<Integer> breakerCooldown;
-        public final CommentedVal<Double> stealerChance;
-        public final CommentedVal<Boolean> breakTileEntities;
-        public final CommentedVal<List<String>> breakingItems;
-        public final CommentedVal<Double> neutralAggressiv;
-        public final CommentedVal<List<String>> autoTargets;
-        public final CommentedVal<Double> difficultyBreak;
-        public final CommentedVal<Double> difficultySteal;
-        public final CommentedVal<Double> flyAIChance;
+        public final CommentedJsonConfig.CommentedVal<List<String>> breakableBlocks;
+        public final CommentedJsonConfig.CommentedVal<Boolean> breakingAsBlacklist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> useBlockBreakSound;
+        public final CommentedJsonConfig.DoubleVal breakerChance;
+        public final CommentedJsonConfig.IntVal breakerInitCooldown;
+        public final CommentedJsonConfig.IntVal breakerCooldown;
+        public final CommentedJsonConfig.DoubleVal stealerChance;
+        public final CommentedJsonConfig.CommentedVal<Boolean> breakTileEntities;
+        public final CommentedJsonConfig.CommentedVal<List<String>> breakingItems;
+        public final CommentedJsonConfig.DoubleVal neutralAggressiv;
+        public final CommentedJsonConfig.CommentedVal<List<String>> autoTargets;
+        public final CommentedJsonConfig.DoubleVal difficultyBreak;
+        public final CommentedJsonConfig.DoubleVal difficultySteal;
+        public final CommentedJsonConfig.DoubleVal flyAIChance;
 
         //Equipment
-        public final CommentedVal<List<String>> equipmentModBlacklist;
-        public final CommentedVal<Boolean> equipmentModWhitelist;
-        public final CommentedVal<List<String>> itemuseBlacklist;
-        public final CommentedVal<Boolean> itemuseWhitelist;
-        public final CommentedVal<List<String>> entityItemConfig;
+        public final CommentedJsonConfig.CommentedVal<List<String>> equipmentModBlacklist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> equipmentModWhitelist;
+        public final CommentedJsonConfig.CommentedVal<List<String>> itemuseBlacklist;
+        public final CommentedJsonConfig.CommentedVal<Boolean> itemuseWhitelist;
+        public final CommentedJsonConfig.CommentedVal<List<String>> entityItemConfig;
 
-        public final CommentedVal<Double> baseEquipChance;
-        public final CommentedVal<Double> baseEquipChanceAdd;
-        public final CommentedVal<Double> diffEquipAdd;
-        public final CommentedVal<Double> baseWeaponChance;
-        public final CommentedVal<Double> diffWeaponChance;
-        public final CommentedVal<Double> baseEnchantChance;
-        public final CommentedVal<Double> diffEnchantAdd;
-        public final CommentedVal<List<String>> enchantCalc;
-        public final CommentedVal<Double> baseItemChance;
-        public final CommentedVal<Double> diffItemChanceAdd;
-        public final CommentedVal<Boolean> shouldDropEquip;
+        public final CommentedJsonConfig.DoubleVal baseEquipChance;
+        public final CommentedJsonConfig.DoubleVal baseEquipChanceAdd;
+        public final CommentedJsonConfig.DoubleVal diffEquipAdd;
+        public final CommentedJsonConfig.DoubleVal baseWeaponChance;
+        public final CommentedJsonConfig.DoubleVal diffWeaponChance;
+        public final CommentedJsonConfig.DoubleVal baseEnchantChance;
+        public final CommentedJsonConfig.DoubleVal diffEnchantAdd;
+        public final CommentedJsonConfig.CommentedVal<List<String>> enchantCalc;
+        public final CommentedJsonConfig.DoubleVal baseItemChance;
+        public final CommentedJsonConfig.DoubleVal diffItemChanceAdd;
+        public final CommentedJsonConfig.CommentedVal<Boolean> shouldDropEquip;
 
         //Attributes
-        public final CommentedVal<Double> healthIncrease;
-        public final CommentedVal<Double> healthMax;
-        public final CommentedVal<Double> roundHP;
-        public final CommentedVal<Double> damageIncrease;
-        public final CommentedVal<Double> damageMax;
-        public final CommentedVal<Double> speedIncrease;
-        public final CommentedVal<Double> speedMax;
-        public final CommentedVal<Double> knockbackIncrease;
-        public final CommentedVal<Double> knockbackMax;
-        public final CommentedVal<Double> magicResIncrease;
-        public final CommentedVal<Double> magicResMax;
-        public final CommentedVal<Double> projectileIncrease;
-        public final CommentedVal<Double> projectileMax;
+        public final CommentedJsonConfig.DoubleVal healthIncrease;
+        public final CommentedJsonConfig.DoubleVal healthMax;
+        public final CommentedJsonConfig.DoubleVal roundHP;
+        public final CommentedJsonConfig.DoubleVal damageIncrease;
+        public final CommentedJsonConfig.DoubleVal damageMax;
+        public final CommentedJsonConfig.DoubleVal speedIncrease;
+        public final CommentedJsonConfig.DoubleVal speedMax;
+        public final CommentedJsonConfig.DoubleVal knockbackIncrease;
+        public final CommentedJsonConfig.DoubleVal knockbackMax;
+        public final CommentedJsonConfig.DoubleVal magicResIncrease;
+        public final CommentedJsonConfig.DoubleVal magicResMax;
+        public final CommentedJsonConfig.DoubleVal projectileIncrease;
+        public final CommentedJsonConfig.DoubleVal projectileMax;
 
-        public CommonConfigVals() {
-            CommentedVal.Builder builder = new CommentedVal.Builder();
+        public CommonConfigVals(CommentedJsonConfig.Builder builder) {
             //builder.comment("With default value every difficulty perk maxes out at difficulty 250").push("general");
             this.enableDifficultyScaling = builder.comment("Disable/Enables the whole difficulty scaling of this mod. Requires a mc restart").define("Enable difficulty scaling", true);
-            this.difficultyDelay = builder.comment("Time in ticks for which the difficulty shouldnt increase at the beginning. One full minecraft day is 24000 ticks").define("Difficulty Delay", 0);
+            this.difficultyDelay = builder.comment("Time in ticks for which the difficulty shouldnt increase at the beginning. One full minecraft day is 24000 ticks").defineInRange("Difficulty Delay", 0, 0, Integer.MAX_VALUE);
             this.increaseHandler = builder.comment("Handles increase in difficulty regarding current difficulty.", "Format is <minimum current difficulty>-<increase every 2400 ticks>", "Example [\"0-0.01\",\"10-0.1\",\"30-0\"]", "So the difficulty increases by 0.01 every 2400 ticks (->0.1 per mc day) till it reaches a difficulty of 10.", "Then it increases by 1 per mc day till it reaches 30 and then stops.").define("Difficulty Increase", Lists.newArrayList("0-0.1, 250-0"));
             this.ignorePlayers = builder.comment("Wether difficulty should only increase with at least one online players or not").define("Ignore Players", false);
             this.shouldPunishTimeSkip = builder.comment("Should punish time skipping with e.g. bed, commands? If false, difficulty will increase by 0.1 regardless of skipped time.").define("Punish Time Skip", true);
@@ -170,20 +177,20 @@ public final class ConfigSpecs {
             this.useScalingHealthMod = builder.comment("Should the scaling health mods difficulty system be used instead of this ones. (Requires scaling health mod)").define("Use Scaling Health Mod", true);
 
             //builder.comment("Settings regarding custom ai for mobs").push("ai");
-            this.breakableBlocks = builder.comment("Whitelist for blocks, which can be actively broken.", BreakableBlocks.use(), "Note: If you include common blocks (like grass blocks) the pathfinding will be a bit strange").define("Block Whitelist", Lists.newArrayList("minecraft:fence_gates", "minecraft:wooden_doors", "c:glass", "c:glass_panes", "minecraft:glass"));
+            this.breakableBlocks = builder.comment("Whitelist for blocks, which can be actively broken. ", BreakableBlocks.use(), "Note: If you include common blocks (like grass blocks) the pathfinding will be a bit strange").define("Block Whitelist", Lists.newArrayList("forge:glass", "forge:glass_panes", "minecraft:fence_gates", "forge:fence_gates", "minecraft:wooden_doors"));
             this.breakingAsBlacklist = builder.comment("Treat Block Whitelist as Blocklist").define("Block as Blacklist", false);
             this.useBlockBreakSound = builder.comment("Use the block breaking sound instead of a knocking sound").define("Sound", false);
-            this.breakerChance = builder.comment("Chance for a mob to be able to break blocks").define("Breaker Chance", 0.3);
-            this.breakerInitCooldown = builder.comment("Initial cooldown for block breaking mobs").define("Breaker Initial Cooldown", 120);
-            this.breakerCooldown = builder.comment("Cooldown for breaking blocks").define("Breaker Cooldown", 20);
-            this.stealerChance = builder.comment("Chance for a mob to be able to steal items").define("Stealer Chance", 0.3);
+            this.breakerChance = builder.comment("Chance for a mob to be able to break blocks").defineInRange("Breaker Chance", 0.3, 0, 1);
+            this.breakerInitCooldown = builder.comment("Initial cooldown for block breaking mobs").defineInRange("Breaker Initial Cooldown", 120, 0, Integer.MAX_VALUE);
+            this.breakerCooldown = builder.comment("Cooldown for breaking blocks").defineInRange("Breaker Cooldown", 20, 0, Integer.MAX_VALUE);
+            this.stealerChance = builder.comment("Chance for a mob to be able to steal items").defineInRange("Stealer Chance", 0.3, 0, 1);
             this.breakingItems = builder.comment("Items which will be given to mobs who can break blocks. Empty list = no items. Syntax: id;weight").define("Breaking items", Lists.newArrayList("minecraft:diamond_pickaxe;1", "minecraft:iron_axe;2"));
             this.breakTileEntities = builder.comment("Should mobs be able to break tile entities? Evaluated before the break list").define("Break Tiles", true);
-            this.neutralAggressiv = builder.comment("Chance for neutral mobs to be aggressive").define("Neutral Aggressive Chance", 0.2);
+            this.neutralAggressiv = builder.comment("Chance for neutral mobs to be aggressive").defineInRange("Neutral Aggressive Chance", 0.2, 0, 1);
             this.autoTargets = builder.comment("List for of pairs containing which mobs auto target others. Syntax is " + MobClassMapConfig.use()).define("Auto Target List", new ArrayList<>());
-            this.difficultyBreak = builder.comment("Difficulty at which mobs are able to break blocks").define("Difficulty Break AI", 0D);
-            this.difficultySteal = builder.comment("Difficulty at which mobs are able to steal items").define("Difficulty Steal AI", 0D);
-            this.flyAIChance = builder.comment("Chance for mobs to be able to ride a parrot").define("Fly Chance", 0.5);
+            this.difficultyBreak = builder.comment("Difficulty at which mobs are able to break blocks").defineInRange("Difficulty Break AI", 0D, 0, Double.MAX_VALUE);
+            this.difficultySteal = builder.comment("Difficulty at which mobs are able to steal items").defineInRange("Difficulty Steal AI", 0D, 0, Double.MAX_VALUE);
+            this.flyAIChance = builder.comment("Chance for mobs to be able to ride a parrot").defineInRange("Fly Chance", 0.5, 0, 1);
 
             //builder.comment("Configs regarding mobs spawning with equipment").push("equipment");
             this.equipmentModBlacklist = builder.comment("Blacklist for mods. Add modid to prevent items from that mod being equipped. (For individual items use the equipment.json)").define("Item Blacklist", new ArrayList<>());
@@ -191,32 +198,32 @@ public final class ConfigSpecs {
             this.itemuseBlacklist = builder.comment("Blacklist for items mobs should never be able to use.", "Use as in using the item similar to players (e.g. shooting bows)").define("Item Use Blacklist", Lists.newArrayList("bigbrain:buckler"));
             this.itemuseWhitelist = builder.comment("Turn the use blacklist into a whitelist").define("Item Use Whitelist", false);
             this.entityItemConfig = builder.comment("Blacklist for specific mobs and items they shouldnt use (e.g. skeletons already use bows)", EntityItemConfig.use()).define("Entity Item Use Blacklist", Config.CommonConfig.entityItemConfig.writeToString());
-            this.baseEquipChance = builder.comment("Base chance that a mob can have one piece of armor").define("Equipment Chance", 0.1);
-            this.baseEquipChanceAdd = builder.comment("Base chance for each additional armor pieces").define("Additional Equipment Chance", 0.3);
-            this.diffEquipAdd = builder.comment("Adds additional x*difficulty% to base equip chance").define("Equipment Addition", 0.3);
-            this.baseWeaponChance = builder.comment("Chance for mobs to have a weapon").define("Weapon Chance", 0.05);
-            this.diffWeaponChance = builder.comment("Adds additional x*difficulty% to base weapon chance").define("Weapon Chance Add", 0.3);
-            this.baseEnchantChance = builder.comment("Base chance for each armor pieces to get enchanted").define("Enchanting Chance", 0.2);
-            this.diffEnchantAdd = builder.comment("Adds additional x*difficulty% to base enchanting chance").define("Enchanting Addition", 0.2);
+            this.baseEquipChance = builder.comment("Base chance that a mob can have one piece of armor").defineInRange("Equipment Chance", 0.1, 0, 1);
+            this.baseEquipChanceAdd = builder.comment("Base chance for each additional armor pieces").defineInRange("Additional Equipment Chance", 0.3, 0, 1);
+            this.diffEquipAdd = builder.comment("Adds additional x*difficulty% to base equip chance").defineInRange("Equipment Addition", 0.3, 0, Double.MAX_VALUE);
+            this.baseWeaponChance = builder.comment("Chance for mobs to have a weapon").defineInRange("Weapon Chance", 0.05, 0, 1);
+            this.diffWeaponChance = builder.comment("Adds additional x*difficulty% to base weapon chance").defineInRange("Weapon Chance Add", 0.3, 0, Double.MAX_VALUE);
+            this.baseEnchantChance = builder.comment("Base chance for each armor pieces to get enchanted").defineInRange("Enchanting Chance", 0.2, 0, 1);
+            this.diffEnchantAdd = builder.comment("Adds additional x*difficulty% to base enchanting chance").defineInRange("Enchanting Addition", 0.2, 0, Double.MAX_VALUE);
             this.enchantCalc = builder.comment("Specify min and max enchanting levels according to difficulty. difficulty-minLevel-maxLevel").define("Enchanting Calc", Lists.newArrayList("0-5-10", "25-5-15", "50-10-17", "100-15-25", "200-20-30", "250-30-35"));
-            this.baseItemChance = builder.comment("Chance for mobs to have an item in offhand").define("Item Equip Chance", 0.05);
-            this.diffItemChanceAdd = builder.comment("Adds additional x*difficulty% to base item chance").define("Item Chance add", 0.2);
+            this.baseItemChance = builder.comment("Chance for mobs to have an item in offhand").defineInRange("Item Equip Chance", 0.05, 0, 1);
+            this.diffItemChanceAdd = builder.comment("Adds additional x*difficulty% to base item chance").defineInRange("Item Chance add", 0.2, 0, Double.MAX_VALUE);
             this.shouldDropEquip = builder.comment("Should mobs drop the armor equipped through this mod? (Other methods e.g. through vanilla is not included)").define("Should drop equipment", false);
 
             //builder.comment("Settings for attribute modifiers").push("attributes");
-            this.healthIncrease = builder.comment("Health will be multiplied by difficulty*0.016*x. Set to 0 to disable").define("Health Increase Multiplier", 1.0);
-            this.healthMax = builder.comment("Health will be multiplied by at maximum this. Set to 0 means no limit").define("Max Health Increase", 5.0);
-            this.roundHP = builder.comment("Round health to the nearest x. Set to 0 to disable").define("Round HP", 0.5);
-            this.damageIncrease = builder.comment("Damage will be multiplied by difficulty*0.008*x. Set to 0 to disable").define("Damage Increase Multiplier", 1.0);
-            this.damageMax = builder.comment("Damage will be multiplied by at maximum this. Set to 0 means no limit").define("Max Damage Increase", 3.0);
-            this.speedIncrease = builder.comment("Speed will be increased by difficulty*0.0008*x. Set to 0 to disable").define("Speed Increase", 1.0);
-            this.speedMax = builder.comment("Maximum increase in speed").define("Max Speed", 0.1);
-            this.knockbackIncrease = builder.comment("Knockback will be increased by difficulty*0.002*x. Set to 0 to disable").define("Knockback Increase", 1.0);
-            this.knockbackMax = builder.comment("Maximum increase in knockback").define("Max Knockback", 0.5);
-            this.magicResIncrease = builder.comment("Magic resistance will be increased by difficulty*0.0016*x. Set to 0 to disable").define("Magic Resistance Increase", 1.0);
-            this.magicResMax = builder.comment("Maximum increase in magic resistance. Magic reduction is percentage").define("Max Magic Resistance", 0.4);
-            this.projectileIncrease = builder.comment("Projectile Damage will be multiplied by 1+difficulty*0.008*x. Set to 0 to disable").define("Projectile Damage Increase", 1.0);
-            this.projectileMax = builder.comment("Projectile damage will be multiplied by maximum of this").define("Max Projectile Damage", 2.0);
+            this.healthIncrease = builder.comment("Health will be multiplied by difficulty*0.016*x. Set to 0 to disable").defineInRange("Health Increase Multiplier", 1.0, 0, Double.MAX_VALUE);
+            this.healthMax = builder.comment("Health will be multiplied by at maximum this. Set to 0 means no limit").defineInRange("Max Health Increase", 5.0, 0, Double.MAX_VALUE);
+            this.roundHP = builder.comment("Round health to the nearest x. Set to 0 to disable").defineInRange("Round HP", 0.5, 0, Double.MAX_VALUE);
+            this.damageIncrease = builder.comment("Damage will be multiplied by difficulty*0.008*x. Set to 0 to disable").defineInRange("Damage Increase Multiplier", 1.0, 0, Double.MAX_VALUE);
+            this.damageMax = builder.comment("Damage will be multiplied by at maximum this. Set to 0 means no limit").defineInRange("Max Damage Increase", 3.0, 0, Double.MAX_VALUE);
+            this.speedIncrease = builder.comment("Speed will be increased by difficulty*0.0008*x. Set to 0 to disable").defineInRange("Speed Increase", 1.0, 0, Double.MAX_VALUE);
+            this.speedMax = builder.comment("Maximum increase in speed").defineInRange("Max Speed", 0.1, 0, 1);
+            this.knockbackIncrease = builder.comment("Knockback will be increased by difficulty*0.002*x. Set to 0 to disable").defineInRange("Knockback Increase", 1.0, 0, Double.MAX_VALUE);
+            this.knockbackMax = builder.comment("Maximum increase in knockback").defineInRange("Max Knockback", 0.5, 0, 1);
+            this.magicResIncrease = builder.comment("Magic resistance will be increased by difficulty*0.0016*x. Set to 0 to disable").defineInRange("Magic Resistance Increase", 1.0, 0, Double.MAX_VALUE);
+            this.magicResMax = builder.comment("Maximum increase in magic resistance. Magic reduction is percentage").defineInRange("Max Magic Resistance", 0.4, 0, 1);
+            this.projectileIncrease = builder.comment("Projectile Damage will be multiplied by 1+difficulty*0.008*x. Set to 0 to disable").defineInRange("Projectile Damage Increase", 1.0, 0, Double.MAX_VALUE);
+            this.projectileMax = builder.comment("Projectile damage will be multiplied by maximum of this").defineInRange("Max Projectile Damage", 2.0, 0, Double.MAX_VALUE);
         }
     }
 }